Refactor price fields in models, services, and DTOs from `Double` to `BigDecimal` for improved precision and consistency.

+++ src/main/java/org/exp/primeapp/models/dto/request/ProductReq.java
@@ -3,6 +3,8 @@ package org.exp.primeapp.models.dto.request;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.Builder;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Builder
@@ -14,7 +16,7 @@ public record ProductReq (
         @NotBlank @Size(max = 512, message = "Brend 512 belgidan oshmasin")
         String brand,
 
-        Double price,
+        BigDecimal price,
         String description,
         Long categoryId,
         List<String> attachmentKeys,
