+++ src/main/java/org/exp/primeapp/models/dto/responce/order/OrdersRes.java
@@ -1,18 +1,17 @@
 package org.exp.primeapp.models.dto.responce.order;
 
 import lombok.Builder;
-import lombok.Value;
-import org.exp.primeapp.models.dto.responce.order.OrderItemRes;
 import org.exp.primeapp.models.dto.responce.user.UserRes;
 import org.exp.primeapp.models.enums.OrderStatus;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Builder
 public record OrdersRes(
         Long id,
         UserRes user,
-        Integer totalPrice,
+        BigDecimal totalPrice,
         OrderStatus status,
         List<OrdersItemRes> ordersItems) {
