+++ src/main/java/org/exp/primeapp/models/dto/responce/order/UserOrderItemRes.java
@@ -2,14 +2,16 @@ package org.exp.primeapp.models.dto.responce.order;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record UserOrderItemRes(
         String name,
         String imageKey,
         String size,
-        Double price,
+        BigDecimal price,
         Integer discount,
         Integer count,
-        Long totalSum
+        BigDecimal totalSum
 ) {
 }
